refactor(obf): make obfToBoard a uniform transform tree, evict path resolver
Inline license and extract transformStrings so obfToBoard returns one
flat literal of named transforms. Move resolveLoadBoardPaths (OBF->OBF
link resolution) out of obf-to-board into board-import beside its
path-index builder, so the path->id map and all its consumers live in
one module and obf/ is purely OBF->Board mapping.
